requirements.txt
mysql-connector-python
requests
tqdm
pandas
matplotlib